Extract shared async API client base class to deduplicate rate limiting and retry logic
USAspending and SAM.gov clients had nearly identical _wait_for_rate_limit()
and _make_request() implementations. Extracted common infrastructure into
BaseAsyncAPIClient, reducing ~250 lines of duplicated code. Also removed
unused retry config attributes (retry_attempts, retry_backoff, retry_multiplier)
that were stored but never referenced by the actual retry decorators.
