+++ sbir_etl/enrichers/sam_gov/client.py
@@ -7,17 +7,15 @@ Supports rate limiting, retry logic, and state management.
 
 from __future__ import annotations
 
-import asyncio
 import os
-from datetime import datetime
 from typing import Any, cast
 
 import httpx
 from loguru import logger
-from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_exponential
 
 from ...config.loader import get_config
-from ...exceptions import APIError, ConfigurationError, RateLimitError
+from ...exceptions import APIError, RateLimitError
+from ..base_client import BaseAsyncAPIClient
 
 
 # Backward compatibility: Alias to central exception classes
@@ -25,8 +23,10 @@ SAMGovAPIError = APIError
 SAMGovRateLimitError = RateLimitError
 
 
-class SAMGovAPIClient:
+class SAMGovAPIClient(BaseAsyncAPIClient):
     """Async client for SAM.gov Entity Information API v3."""
+
+    api_name = "sam_gov"
 
     def __init__(
         self,
@@ -40,6 +40,8 @@ class SAMGovAPIClient:
             config: Optional configuration override. If None, loads from get_config()
             http_client: Optional pre-configured HTTPX client (useful for tests)
         """
+        super().__init__()
+
         if config is None:
             cfg = get_config()
             self.api_config = cfg.enrichment.sam_gov
@@ -50,8 +52,6 @@ class SAMGovAPIClient:
             self.api_config.get("base_url", "https://api.sam.gov/entity-information/v3")
         )
         self.timeout = cast(int, self.api_config.get("timeout_seconds", 30))
-        self.retry_attempts = cast(int, self.api_config.get("retry_attempts", 3))
-        self.retry_backoff = cast(float, self.api_config.get("retry_backoff_seconds", 1.0))
         self.rate_limit_per_minute = cast(int, self.api_config.get("rate_limit_per_minute", 60))
 
         # Get API key from environment
@@ -63,10 +63,6 @@ class SAMGovAPIClient:
                 "API calls will fail without authentication."
             )
 
-        # Rate limiting state
-        self.request_times: list[datetime] = []
-        self._rate_limit_lock = asyncio.Lock()
-
         self._client = http_client or httpx.AsyncClient(timeout=self.timeout)
 
         logger.info(
@@ -74,109 +70,12 @@ class SAMGovAPIClient:
             f"rate_limit={self.rate_limit_per_minute}/min"
         )
 
-    async def aclose(self) -> None:
-        """Close the underlying HTTP client."""
-        await self._client.aclose()
-
-    async def _wait_for_rate_limit(self) -> None:
-        """Wait if rate limit would be exceeded."""
-        async with self._rate_limit_lock:
-            now = datetime.now()
-            self.request_times = [t for t in self.request_times if (now - t).total_seconds() < 60]
-
-            if len(self.request_times) >= self.rate_limit_per_minute:
-                oldest = min(self.request_times)
-                wait_seconds = 60 - (now - oldest).total_seconds() + 1
-                if wait_seconds > 0:
-                    logger.debug(f"Rate limit reached, waiting {wait_seconds:.1f} seconds")
-                    await asyncio.sleep(wait_seconds)
-
-            self.request_times.append(datetime.now())
-
-    async def _make_request(
-        self,
-        method: str,
-        endpoint: str,
-        params: dict[str, Any] | None = None,
-        *,
-        headers: dict[str, str] | None = None,
-    ) -> dict[str, Any]:
-        """Make an HTTP request to SAM.gov API with retry logic.
-
-        Args:
-            method: HTTP method (GET, POST)
-            endpoint: API endpoint path (relative to base_url)
-            params: Query parameters
-            headers: Additional headers
-
-        Returns:
-            JSON response as dict
-
-        Raises:
-            SAMGovAPIError: If request fails
-            SAMGovRateLimitError: If rate limit exceeded
-        """
-
-        # Inner function that does the actual request (will be wrapped by retry decorator)
-        @retry(
-            stop=stop_after_attempt(3),
-            wait=wait_exponential(multiplier=2.0, min=1, max=30),
-            retry=retry_if_exception_type((httpx.HTTPError, httpx.TimeoutException)),
-            reraise=True,
-        )
-        async def _do_request() -> dict[str, Any]:
-            await self._wait_for_rate_limit()
-
-            base_url_str = str(self.base_url)
-            url = f"{base_url_str.rstrip('/')}/{endpoint.lstrip('/')}"
-            default_headers = {
-                "Accept": "application/json",
-                "User-Agent": "SBIR-Analytics/0.1.0",
-            }
-            if self.api_key:
-                default_headers["X-Api-Key"] = self.api_key
-
-            if headers:
-                default_headers.update(headers)
-
-            try:
-                if method.upper() == "GET":
-                    response = await self._client.get(url, params=params, headers=default_headers)
-                elif method.upper() == "POST":
-                    response = await self._client.post(url, json=params, headers=default_headers)
-                else:
-                    raise ConfigurationError(f"Unsupported HTTP method: {method}")
-
-                response.raise_for_status()
-
-                # Handle rate limit errors
-                if response.status_code == 429:
-                    raise RateLimitError(
-                        f"SAM.gov API rate limit exceeded: {response.text}",
-                        status_code=429,
-                        response_text=response.text,
-                    )
-
-                return response.json()
-
-            except httpx.HTTPStatusError as e:
-                if e.response.status_code == 429:
-                    raise RateLimitError(
-                        f"SAM.gov API rate limit exceeded: {e.response.text}",
-                        status_code=429,
-                        response_text=e.response.text,
-                    )
-                raise APIError(
-                    f"SAM.gov API request failed: {e.response.status_code} - {e.response.text}",
-                    status_code=e.response.status_code,
-                    response_text=e.response.text,
-                )
-            except httpx.TimeoutException as e:
-                raise APIError(f"SAM.gov API request timed out: {e}")
-            except httpx.RequestError as e:
-                raise APIError(f"SAM.gov API request error: {e}")
-
-        return await _do_request()
+    def _build_headers(self) -> dict[str, str]:
+        """Build default headers including API key if available."""
+        headers = super()._build_headers()
+        if self.api_key:
+            headers["X-Api-Key"] = self.api_key
+        return headers
 
     async def get_entity_by_uei(self, uei: str) -> dict[str, Any] | None:
         """Get entity data by UEI (Unique Entity Identifier).
